+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+-  `Connection.authenticate_oidc()`: add argument to set maximum device code flow poll time
+
 ### Changed
 
 ### Removed